Initial Commit
First upload.

Loads a JSON file containing the inti settings, turns it into a string
and then shoots it via serial to the tinyG. A GUI is made with the
ControlP5 library to provide a text area to show the communication
hisotry and a text field to allow for GCode or tinyG commands injection.