Buffer overflow control and init
Increased the time between lines of code being dumped by dumpFile() when
the file is not init.json. The buffer on the tinyG was overflowing
causing a crash.

According to the Wiki, the tinyG's motion planner takes about 25ms to
process a MOVE command (G1, G0, etc), when it only takes <10ms to
process a NON MOVE command (setting).

It can only take about 500 chars before it overflows, so sending them at
a slower pace prevents this from happening... hopefully.

The ToDo is a personal list of things to implement.