Added LogFile cleaner
Added functions to make the log files human-readable. Click on the third
button to start the script. A file loader will pop to let you choose the
right file, then another will pop to let you choose a file location. Hit
save and it will clean the file.

Also cleaned the way the terminal shows some of the input from the
tinyG. It was getting really messy.