Major changes
This version incorporates DinBoy's implementation of the queue. The
machine will move slower than in previous versions, but now the app
doesn't freeze until all the file has been sent.

It also includes a few different methods for homing the DIWire and for
using the smartpin (using "measure")

Now it also allows to indicate how many times you want to dump a file
and keeps track of the progress. For this to work the file's last line
must be the word "END"